+++ src/main/java/gift/service/CategoryService.java
@@ -1,15 +1,53 @@
 package gift.service;
 
+import gift.database.repository.JpaCategoryRepository;
 import gift.dto.CategoryRequest;
 import gift.dto.CategoryResponse;
+import gift.model.Category;
+import jakarta.transaction.Transactional;
 import java.util.List;
+import org.springframework.stereotype.Service;
 
-public interface CategoryService {
-    List<CategoryResponse> readAll();
+@Service
+@Transactional
+public class CategoryService{
 
-    void create( CategoryRequest categoryRequest);
+    private final JpaCategoryRepository jpaCategoryRepository;
 
-    void update(long id, CategoryRequest categoryRequest);
+    public CategoryService(JpaCategoryRepository jpaCategoryRepository) {
+        this.jpaCategoryRepository = jpaCategoryRepository;
+    }
 
-    void delete(long id);
+    public List<CategoryResponse> readAll() {
+        return jpaCategoryRepository.findAll()
+            .stream()
+            .map(CategoryResponse::new)
+            .toList();
+
+    }
+
+    public void create(CategoryRequest categoryRequest) {
+        Category category = getCategory(null, categoryRequest);
+
+        jpaCategoryRepository.save(category);
+    }
+
+
+    public void update(long id, CategoryRequest categoryRequest) {
+        Category category = getCategory(id, categoryRequest);
+        jpaCategoryRepository.save(category);
+    }
+
+    public void delete(long id) {
+        jpaCategoryRepository.deleteById(id);
+    }
+
+    private static Category getCategory(Long id, CategoryRequest categoryRequest) {
+        return new Category(id,
+            categoryRequest.getName(),
+            categoryRequest.getColor(),
+            categoryRequest.getDescription(),
+            categoryRequest.getImageUrl());
+    }
+
 }
