+++ src/main/java/gift/service/WishListService.java
@@ -1,17 +1,76 @@
 package gift.service;
 
+
+import gift.database.repository.JpaMemberRepository;
+import gift.database.repository.JpaProductRepository;
+import gift.database.repository.JpaWishRepository;
 import gift.dto.WishListResponse;
+import gift.exceptionAdvisor.exceptions.GiftNotFoundException;
+import gift.model.Member;
+import gift.model.Product;
+import gift.model.Wish;
+import java.util.Map;
+import java.util.stream.Collectors;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface WishListService {
+@Service
+@Transactional
+public class WishListService {
 
-    void addProduct(long memberId, long productId);
+    private final JpaMemberRepository jpaMemberRepository;
+    private final JpaWishRepository jpaWishRepository;
+    private final JpaProductRepository jpaProductRepository;
 
-    void deleteProduct(long memberId, long productId);
+    public WishListService(JpaMemberRepository jpaMemberRepository,
+        JpaWishRepository jpaWishRepository, JpaProductRepository jpaProductRepository) {
+        this.jpaMemberRepository = jpaMemberRepository;
+        this.jpaWishRepository = jpaWishRepository;
+        this.jpaProductRepository = jpaProductRepository;
+    }
 
-    void updateProduct(long memberId, long productId, int productValue);
+    public void addProduct(long memberId, long productId) {
+        Member member = jpaMemberRepository.findById(memberId)
+            .orElseThrow(()->new GiftNotFoundException("회원이 존재하지 않습니다."));
+        Product product = jpaProductRepository.findById(productId).orElseThrow();
+        Wish wish = new Wish(member, product);
+        member.addWish(wish);
+        jpaWishRepository.save(wish);
+    }
 
-    WishListResponse getWishList(long memberId);
+    public void deleteProduct(long memberId, long productId) {
+        Member member = jpaMemberRepository.findById(memberId)
+            .orElseThrow(()-> new GiftNotFoundException("회원이 존재하지 않습니다."));
+        Product product = jpaProductRepository.findById(productId)
+            .orElseThrow(()-> new GiftNotFoundException("상품이 존재하지 않습니다."));
+        Wish wish = jpaWishRepository.findByMemberIdAndProductId(memberId, productId)
+            .orElseThrow(()-> new GiftNotFoundException("wish가 존재하지 않습니다."));
+        member.delWish(wish);
+        jpaWishRepository.delete(wish);
+    }
 
-    WishListResponse getWishListPage(long memberId, int pageNumber, int pageSize);
+    public void updateProduct(long memberId, long productId, int productValue) {
+        Wish wish = jpaWishRepository.findByMemberIdAndProductId(memberId, productId)
+            .orElseThrow(()-> new GiftNotFoundException("wish가 존재하지 않습니다."));
 
+        wish.productCountUpdate(productValue);
+
+    }
+
+    public WishListResponse getWishList(long memberId) {
+        Member member = jpaMemberRepository.findById(memberId)
+            .orElseThrow(()-> new GiftNotFoundException("회원이 존재하지 않습니다."));
+        Map<String, Integer> wishList = member.getWishList().stream()
+            .collect(Collectors.toMap(Wish::getProductName, Wish::getProductCount));
+        return new WishListResponse(member.getId(), wishList);
+    }
+
+    public WishListResponse getWishListPage(long memberId, int pageNumber, int pageSize) {
+        Pageable pageable = PageRequest.of(pageNumber, pageSize);
+        Map<String, Integer> wishlist = jpaWishRepository.findByMemberId(memberId, pageable)
+            .stream().collect(Collectors.toMap(Wish::getProductName, Wish::getProductCount));
+        return new WishListResponse(memberId, wishlist);
+    }
 }
